requirements.txt
azure-storage-queue>=1.1.0,<1.1.99